use act_as_list in artwork model

+++ db/migrate/20151008175425_change_data_type_for_position.rb
@@ -1,7 +1,7 @@
 class ChangeDataTypeForPosition < ActiveRecord::Migration
   def change
     change_table :artworks do |t|
-      t.change :position, :float
+      t.change :position, :integer
     end
   end
 end
